Moved the `polar_coordinates` function
Added some tests to check consistency of spherical coordinates
